- Refactoring deja en basic_logger lo minimo indispensable
