+++ php/nucleo/lib/log_handlers/toba_basic_logger.php
@@ -28,19 +28,13 @@ if (!defined('apex_log_error_log_nivel')) {
 
 trait toba_basic_logger 
 {	
-	public static $separador = "-o-o-o-o-o-";
-	public static $fin_encabezado = "==========";
-	public static $limite_mensaje = 100000; //100 KB
-
 	protected $ref_niveles = array("EMERGENCY" , "ALERT", "CRITICAL", "ERROR", "WARNING", "NOTICE", "INFO", "DEBUG");	
 	protected $mensajes = array();
 	protected $niveles = array();
 	protected $proyectos = array();
 	protected $proyecto_actual;
 	
-	protected $proximo = 0;
 	protected $nivel_maximo = 7;
-	protected $activo = true;
 	
 	protected $es_php_compatible = true;
 	
@@ -54,15 +48,6 @@ trait toba_basic_logger
 			}
 		}
 		return 'toba';
-	}
-	
-	public function get_usuario_actual()
-	{
-		if( php_sapi_name() === 'cli' ) {
-			return null;
-		} else {
-			return toba::usuario()->get_id();
-		}
 	}
 	
 	/**
@@ -115,54 +100,6 @@ trait toba_basic_logger
 	{
 		return $this->nivel_maximo;
 	}	
-	
-	//------------------------------------------------------------------
-	//---- Manejo de las fuentes de log
-	//------------------------------------------------------------------	
-	public function directorio_logs()
-	{
-		if (! isset($this->dir_logs)) {
-			$id_instancia = toba_instancia::get_id();
-			$this->dir_logs = toba_nucleo::toba_instalacion_dir()."/i__$id_instancia/p__{$this->proyecto_actual}/logs";
-		}
-		return $this->dir_logs;
-	}	
-	
-	public function set_directorio_logs($dir)
-	{
-		$this->dir_logs = $dir;	
-	}
-	
-	//-----------------------------------------------------------------------------------------------------------//	
-	//					METODOS AUXILIARES
-	//-----------------------------------------------------------------------------------------------------------//
-	protected function truncar_msg($msg)
-	{
-		if (strlen($msg) > self::$limite_mensaje) {
-			$msg = substr($msg, 0, self::$limite_mensaje).
-					"..TEXTO CORTADO POR EXCEDER EL LIMITE DE ".
-					self::$limite_mensaje.
-					" bytes";
-		}
-		return $msg;
-	}
-	
-	protected function registrar_msg_cli($msg, $nivel)
-	{
-		if (PHP_SAPI != 'cli' && apex_log_error_log && $nivel <= apex_log_error_log_nivel) {
-			$error_log_max = ini_get("log_errors_max_len");
-			if (! isset($error_log_max) || !is_numeric($error_log_max) || strlen($error_log_max) <= 1) {
-				$error_log_max = 1024;
-			}
-			$error_log_extra = "...SIGUE...";
-			$msg_error_log = $msg;
-			if (strlen($msg_error_log) > $error_log_max) {
-				$msg_error_log = substr($msg_error_log, 0 , $error_log_max - strlen($error_log_extra));
-				$msg_error_log .= $error_log_extra;
-			}
-			error_log($msg_error_log);
-		}		
-	}
 	
 	/**
 	 * @ignore 
@@ -324,76 +261,6 @@ trait toba_basic_logger
 		return ((1 << ($nivel + 1)) - 1);
 	}
 		
-	protected function ciclar_archivos_logs($path, $archivo)
-	{
-		if (apex_log_archivo_backup_cant == 0) {
-			//Si es un unico archivo hay que borrarlo
-			unlink($path."/".$archivo);
-			return;
-		}
-		//Encuentra los archivos
-		$patron = "/$archivo\.([0-9]+)/";
-		$archivos = toba_manejador_archivos::get_archivos_directorio($path, $patron);
-		sort($archivos);
-
-		//¿Cual es el numero de cada uno?
-		$ultimo = 0;
-		$arch_ordenados = array();
-		foreach ($archivos as $arch_actual) {
-			$version = array();
-			preg_match($patron, $arch_actual, $version);
-			if (! empty($version) && count($version) > 1) {
-				$pos = $version[1];
-				$arch_ordenados[$pos] = $arch_actual;
-				if ($pos > $ultimo) {
-					$ultimo = $pos;
-				}
-			}
-		}
-		//Se determina el siguiente numero
-		$sig = $ultimo + 1;
-		
-		//¿Hay que purgar algunos?
-		$this->purgar_archivos_viejos($arch_ordenados);
-	
-		//Se procede a mover el archivo actual
-		$path_completo = $path . "/" . $archivo;
-		if (apex_log_archivo_backup_compr) {
-			//Se comprime
-			$nuevo = $path_completo . ".$sig.gz";
-			toba_manejador_archivos::comprimir_archivo($path_completo, 5, $nuevo);
-			unlink($path_completo);
-		} else {
-			$nuevo = $path_completo . ".$sig";
-			rename($path_completo, $nuevo);
-		}
-	}
-
-	protected function purgar_archivos_viejos($lista_archivos)
-	{
-		//¿Hay que purgar algunos?
-		$puede_purgar = (apex_log_archivo_backup_cant != -1);
-		$a_purgar = count($lista_archivos) - (apex_log_archivo_backup_cant -1); 			//Se dejan solo N-1 archivos
-		if ($puede_purgar && $a_purgar > 0) {
-			ksort($lista_archivos);
-			reset($lista_archivos);
-			 do {
-				unlink(current($lista_archivos));
-				$a_purgar--;
-				next($lista_archivos);
-			} while ($a_purgar > 0);
-		}		
-	}
-	
-	protected function anexar_a_archivo($texto, $archivo)
-	{
-		$res = file_put_contents($archivo, "$texto\r\n", FILE_APPEND);
-		if ($res === FALSE) {
-			throw new toba_error("Imposible guardar el archivo de log '$archivo'. Chequee los permisos de escritura del usuario apache sobre esta carpeta/archivo");
-		}
-	}	
-	
-	
 	//------------------------------------------------------------------
 	//------ Entradas para los distintos tipos de error
 	//------------------------------------------------------------------
